requirements.txt
matplotlib>=3.5.0
numpy>=1.20.0
pillow>=8.0.0
torch>=1.12.0
transformers>=4.20.0
tqdm>=4.60.0
accelerate>=0.20.0
